chore: (WIP) phoneNumber validations, convertions and exceptions
