--- src/main/java/buildings/controllers/implementations/BuildingController.java
@@ -1,5 +1,0 @@
-package buildings.controllers.implementations;
-
-public class BuildingController {
-
-}
