--- src/main/java/buildings/controllers/responses/BuildingResponse.java
@@ -1,5 +1,0 @@
-package buildings.controllers.responses;
-
-public record BuildingResponse() {
-
-}
