--- src/main/java/buildings/data/models/BuildingType.java
@@ -1,21 +1,0 @@
-package buildings.data.models;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.Id;
-import lombok.*;
-
-import java.util.UUID;
-
-@Entity(name="building_types")
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-@Data
-public class BuildingType {
-    @Id
-    private UUID buildingTypeId = UUID.randomUUID();
-
-    @NonNull
-    private String name;
-
-}
