--- src/main/java/shared/data/models/Locatable.java
@@ -1,18 +1,0 @@
-package shared.data.models;
-
-import jakarta.persistence.Column;
-
-import jakarta.persistence.MappedSuperclass;
-
-
-
-@MappedSuperclass
-public abstract class Locatable {
-
-    @Column(name="latitude")
-    private Double latitude;
-
-    @Column(name="longitude")
-    private Double longitude;
-
-}
