--- src/main/java/users/data/models/User.java
@@ -1,4 +1,0 @@
-package users.data.models;
-
-public class User {
-}
